pyproject.toml
[tool.poetry]
name = "andoird touch replay"
version = "0.1.0"
description = ""
authors = ["IsmaelBena <ismael.benadjal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
keyboard = "^0.13.5"
pyyaml = "^6.0.2"
kivy = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
